requirements.txt
fasttext==0.9.2
h5py==2.10.0
icrawler==0.6.3
imageio==2.6.1
newspaper3k==0.2.8
numpy==1.18.4
opencv-python
PyYAML==5.4.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
spacy==2.2.4
torch==1.4.0
torchvision==0.5.0
tqdm==4.46.0